backend/requirements.txt
# =============================================================
#  PhishGuard AI — Python Dependencies
# =============================================================
#
#  HOW TO USE:
#    After activating your virtual environment, run:
#      pip install -r requirements.txt
#
#  WHY a requirements.txt?
#    When you share this project, teammates need to know which
#    packages to install. This file is that recipe list.
#    The "==" pins the EXACT version — so everyone gets
#    identical behaviour. Without pinning, "pip install flask"
#    today and tomorrow could give different versions.
#
#  SETUP ORDER (run these once, inside /backend):
#    1. python -m venv venv          ← create isolated environment
#    2. source venv/bin/activate     ← Mac/Linux
#       venv\Scripts\activate        ← Windows
#    3. pip install -r requirements.txt
#
#  You'll know venv is active when your prompt shows (venv):
#    (venv) C:\phishguard\backend>
# =============================================================


# ── WEB FRAMEWORK ─────────────────────────────────────────────
# Flask     : core web framework. Turns Python functions into
#             HTTP endpoints a browser can call.
# Flask-CORS: browsers block JS on port 5500 calling an API on
#             port 5000 by default. This adds the headers that
#             tell the browser "cross-origin is allowed here".
Flask==3.0.3
Flask-CORS==4.0.1


# ── DATABASE ──────────────────────────────────────────────────
# Flask-SQLAlchemy: ORM — lets you write Python classes instead
#             of raw SQL. A class = a table, an instance = a row.
# Flask-Migrate   : tracks schema changes over time (like Git
#             for your database). Add a column → create a
#             migration → apply it without losing existing data.
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7


# ── AUTHENTICATION ────────────────────────────────────────────
# PyJWT  : creates and verifies JSON Web Tokens (login tokens).
#          After login, the server gives the user a signed token.
#          They send it with every request to prove who they are.
# bcrypt : one-way password hashing. NEVER store plain passwords.
#          Even if your DB leaks, bcrypt hashes are useless to
#          attackers. It is intentionally slow to resist brute-force.
PyJWT==2.8.0
bcrypt==4.1.3


# ── MACHINE LEARNING (used from Phase 2 onward) ───────────────
# scikit-learn : TF-IDF vectorizer, Random Forest, accuracy tools.
# numpy        : fast numeric arrays. scikit-learn needs this.
# pandas       : spreadsheet-style data handling. Good for CSVs.
# joblib       : save/load trained models to disk so we train
#                once and serve instantly on every restart.
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2


# ── TEXT & EMAIL PROCESSING (Phase 2) ────────────────────────
# nltk         : tokenisation, stopwords, stemming.
# regex        : more powerful than Python's built-in re module.
# mail-parser  : parses raw .eml files and extracts headers,
#                body, and attachments cleanly.
nltk==3.8.1
regex==2024.5.15
mail-parser==3.15.0


# ── HTTP / INTEGRATIONS (Phase 6) ────────────────────────────
# requests : standard Python HTTP client used for Slack
#            webhooks and any external API calls.
requests==2.32.3


# ── PDF REPORTS (Phase 6) ────────────────────────────────────
reportlab==4.2.2


# ── TESTING (Phase 6) ────────────────────────────────────────
# pytest       : discovers test_*.py files and runs them.
# pytest-flask : Flask helpers for testing routes without a
#                real running server.
# Faker        : generates realistic fake names, emails, text.
pytest==8.2.2
pytest-flask==1.3.0
Faker==25.9.2


# ── UTILITIES ─────────────────────────────────────────────────
# python-dotenv : reads your .env file into os.environ so
#                 secret keys never have to live in code.
python-dotenv==1.0.1